libproxy: add a PAC/WPAD handling D-Bus service
libproxy can deal with PAC and WPAD; other proxy resolvers may not be
able to, so let them use libproxy to do it.

Having it out-of-process is good because the only way to force
libproxy to do exactly what we want is to fiddle with environment
variables, and we don't want those to leak into the environment.

https://bugzilla.gnome.org/show_bug.cgi?id=644211