gnutls: Implement GTlsDatabase and related objects
The database is an abstract object implemented by the various TLS
backends, which is used by GTlsConnection to lookup certificates
and keys, as well as verify certificate chains.

Here we implement a file based database to lookup anchor certificates
for use with GnuTLS.

Also hookup the various bits for GTlsInteraction, but don't actually
use that yet.

https://bugzilla.gnome.org/show_bug.cgi?id=636572