tls/tests: fix up leak checking
You can't use !G_IS_FOO(object) to test that an object has been
destroyed, since (a) it may not work if the memory hasn't been
cleared, and (b) it might crash if it has. Instead, use weak pointers
to test that the object has been destroyed when we expect it to be.