tls/tests/connection: fix some race conditions
Close the client-side connection cleanly, to avoid a race condition
where the server side might see an unclean close.

In teardown_connection(), loop waiting for objects to be destroyed
rather than just asserting that the ref we hold is the last one; in
some cases, a GTask in another thread may still be holding a reference
on one of the objects for an instant after teardown_connection() is
called.

https://bugzilla.gnome.org/show_bug.cgi?id=693720