proxy/tests: fix to work when neither proxy module is available
automake ends up setting TEST_PROGS to " " in this case, but
Makefile.decl was assuming that if TEST_PROGS is not empty then it
contains actual tests, and then ended up bailing out by passing an no
tests to gtester. Fix it to use $(strip) to get the right result.

Also, drive-by automake warning fix in proxy/tests/Makefile.am

https://bugzilla.gnome.org/show_bug.cgi?id=700628