Revert "gnutls: require gnutls 3.0"
There are potential licensing issues with gnutls 3.x, which
transitively requires gmp, whose latest versions are LGPLv3.

Allow compiling with 2.x again, but bump up the requirement to 2.12,
so we've got all the newer APIs to replace the ones that are
deprecated in 3.x.

This may be revisited in the future, and some new code added to this
module will probably be gnutls-3.x-only.

https://bugzilla.gnome.org/show_bug.cgi?id=704503