GTlsConnection: avoid non-literal format strings
The error handling in GTlsConnection involved translated format strings
that were then passed to another function where the arguments were
applied.  This prevents the C compiler from enforcing restrictions on
the format string and runs afoul of -Werror=format-nonliteral, breaking
the build under clang.

Fix it up so that the error handling function takes the format string
along with the argument and calls g_error_new_valist().  This allows the
compiler to check that we're getting it right.

https://bugzilla.gnome.org/show_bug.cgi?id=720081