gnutls: Prevent use of a destroyed GSource in a callback
It’s possible for the GTlsConnectionGnutlsSource to be destroyed in its
user-provided callback function. If gnutls_source_sync() is called after
this, it ends up attempting to modify the GSource after it’s been
destroyed, which is disallowed.

Bail out of gnutls_source_sync() early if the source has already been
destroyed.

https://bugzilla.gnome.org/show_bug.cgi?id=723774