test: add script to re-generate test certificates and generate new ones
Added a create-files.sh script under tls/tests/files that re-generates
all test certificates in the case that it's needed. It will also print
instructions about a few manual changes that need to be made.

This patch also creates new certificates with a new CA. The new CA
private key is also being committed so it's not lost.

https://bugzilla.gnome.org/show_bug.cgi?id=733365