gnutls: Ensure all private members of GTlsConnectionGnutls are freed
Just to be on the safe side, ensure every private member of
GTlsConnectionGnutls is freed when the object itself is finalised.

implicit_handshake must always be NULL when finalize() is reached, as it
holds a reference to the GTlsConnectionGnutls as its source object.
However, it’s better to clear the member anyway, just in case this
behaviour changes in future.

https://bugzilla.gnome.org/show_bug.cgi?id=736809