gnutls: Fix leaking GTask on TLS close after implicit handshake
If a GTlsConnectionGnutls is created and an implicit handshake operation
is started on it (e.g. via a read), if the handshake doesn’t complete
and close() is called on the connection without any other operations
being called first, the GTlsConnectionGnutls.implicit_handshake GTask
will be leaked. Since it holds a reference to the GTlsConnectionGnutls
connection as a whole, that object will be leaked too in a reference
cycle.

Fix this by explicitly clearing implicit_handshake when claiming
OP_CLOSE, instead of ignoring it. Any errors from the handshake process
are ignored as before, however.

https://bugzilla.gnome.org/show_bug.cgi?id=736809