gnutls: Add support for DTLS to the GnuTLS backend
Implement DTLS support in the GnuTLS backend by adding it to
GTlsConnectionGnutls, so that it supports both TLS and DTLS (despite its
name). This means the GTlsBackendGnutls returns the same type for client
and server connection types — the protocol to use is determined by
whether the GTlsConnection:base-io-stream or
GDtlsConnection:base-socket property is set on the
GTlsConnectionGnutls. Exactly one of the two must be set.

This makes the following internal API changes:
 • Implement GDtlsClientConnection on GTlsClientConnectionGnutls
 • Implement GDtlsServerConnection on GTlsServerConnectionGnutls
 • Implement GDatagramBased on GTlsConnectionGnutls
 • Implement GDtlsConnection on GTlsConnectionGnutls

The patch doesn’t support vectored I/O; support will follow in a later
commit.

It includes contributions by Olivier Crête
<olivier.crete@collabora.com>.

https://bugzilla.gnome.org/show_bug.cgi?id=697908