Revert "gnutls: use gnutls's verification code for GTlsFileDatabaseGnutls"
This reverts commit 55c61d1c2de20af86adf711e7b6e10787077420a.

The commit implicitly bumps up the dependency of GnuTLS for
glib-networking to be ≥ 3.3, because of the use of the
gnutls_x509_trust_list_verify_crt2() function.

Glib-networking still depends on GnuTLS 3.0; at the very least, this
commit should also bump up the dependency inside configure.ac — but both
JHBuild and Continuous provide older versions.

https://bugzilla.gnome.org/show_bug.cgi?id=761713