Visual Studio builds: Move Projects to win32/
This enables one to go down one less level into the source tree to reach
the project files, as well as getting the autotools modules (Makefile.msvcproj,
Makefile-newvs.am) to be in sync with the ones in GLib master.

This also adds support for Visual Studio 2017, by updating the autotools scripts
to copy the 2010 projects and updating items as necessary to obtain the 2017
projects.

Note that the toolset version string format changed for Visual Studio 2017, so
a custom toolset version string is allowed and used if provided, otherwise it
will be generated as before.

The modules built with Visual Studio 2017 should work without problems with
the 2015-built ones, as 2017 aims to be compatible with 2015 on the CRT level.