gnutls: Stop using %LATEST_RECORD_VERSION in priority string
This was added after POODLE to deal with broken servers that conflated
TLS record version with protocol version and started blocking clients
that used the SSLv3 record version. Now that SSLv3 is no longer enabled
by WebKit or newer versions of GnuTLS, there is no longer any reason
to keep doing this, and it's breaking interoperability with other broken
servers. Remove it.

This also adds a comment to clarify the confusing duplication of
%COMPAT in the fallback priority string.

https://bugzilla.gnome.org/show_bug.cgi?id=782218