Ensure tests run against the GIO modules in the build root
Currently, our tests run against the GIO modules that are installed on
the system. So if we 'ninja test' before running 'ninja install', we'll
be testing the wrong code. And this is actually the default case when
running 'jhbuild buildone -n --check', so that's bad.

This was actually broken in the Autotools build, too. We just never
noticed before. I discovered this after testing a build without TLS
support: after reenabling TLS support and rebuilding glib-networking,
but before installing it, all the TLS tests failed.

https://bugzilla.gnome.org/show_bug.cgi?id=789171