Upgrade license to LGPLv2.1+ and standardize license headers
It's mildly annoying to read the current license headers, which grant
permission to use the library under the terms of the GNU Lesser General
Public License version 2 or later, but there is no such license (as it
was the GNU Library General Public License way back then). So upgrade
from version 2 to version 2.1. This is what we have specified in the
meson.build anyway.

This should almost certainly have no effect on anything.