meson: Improve Visual Studio build experience
The OpenSSL build system does not generate pkg-config files for us, and
crafting pkg-config files by hand may be tedious and error-prone, so add
a fallback to check for the OpenSSL headers and libraries manually for
Visual Studio builds.  As the .lib files for OpenSSL 1.1.x and 1.0.x
(and earlier) are different, we need to determine the correct pair of
.lib files that we want to link to later.

Also, to clean things up a bit, just force-include
msvc_recommended_pragmas.h instead of checking for the warning cflags on
Visual Studio, as they exist on all the Visual Studio versions that we
support.

https://bugzilla.gnome.org/show_bug.cgi?id=790886