Fallback tests should not expect SSLv3 to be available
Our SSLv3 fallback test does not work if GnuTLS is built without SSLv3
support.

Currently we run the test twice, with two different priorities:

 #define PRIORITY_SSL_FALLBACK "NORMAL:+VERS-SSL3.0"
 #define PRIORITY_TLS_FALLBACK "NORMAL:+VERS-TLS-ALL:-VERS-SSL3.0"

PRIORITY_SSL_FALLBACK is guaranteed to be broken with newer GnuTLS, and
PRIORITY_TLS_FALLBACK is effectively the default case, so neither is
interesting. We should just not set custom priority strings for the
test. That will avoid the problem entirely, and allow us to simplify
the test by getting rid of the subprocess stuff. The priority strings
are not supposed to be API anyway; our main concern is to make sure that
GTlsClientConnection:use-ssl3 works as "expected" (to force use of the
lowest-supported protocol version, which should be TLS 1.0).

https://bugzilla.gnome.org/show_bug.cgi?id=782853