Use G_DECLARE_[FINAL,DERIVABLE]_TYPE
This means we eliminate the priv pointer from final classes, and access
it via the _get_instance_private() getter for derivable classes. It's a
mechanical transformation, except in GTlsCertificateGnutls where a
couple loops need adjusted to make it easier to get the priv struct of
other GTlsCertificateGnutls objects.

This is considered the modern best-practice, and is possible now that we
finally depend on a recent version of GLib.