meson: Improve GIO modules directory handling
GIO modules are usually installed in a system directory that is
checked from the `gio-2.0.pc` file. However this might cause issues
to users without permissions to write in those system directories.

A new option has been added to allow anyone to choose a writable
directory. If this option is let empty it will fallback to the
previous behaviour by checking the information from the `gio-2.0.pc`
file.

The meson variable has also been renamed to prevent any confusion.

https://bugzilla.gnome.org/show_bug.cgi?id=793311