proxy: Use XDG_CURRENT_DESKTOP to determine which proxy module to load
This will somewhat reduce the odds that that we'll load libproxy in the
gnome-shell greeter process, causing libproxy's old version of mozjs to
be loaded into the same process as gjs's newer version, causing the
login screen to crash in a spectacular loop of sadness, causing
coredumpctl to thrash the system... yeah.

This is already fixed in gnome-session by
https://git.gnome.org/browse/gnome-session/commit/?id=5f920253dfb2c0f2c4b742690f306cfea2ac6613
but I figure I should do my civic duty here to make it easier to remove
these environment variables in the future.

https://bugzilla.gnome.org/show_bug.cgi?id=790387