Fix crash when setting client cert without private key
If, if g_tls_client_connection_gnutls_retrieve_function, we return to
GnuTLS a gnutls_retr2_st with positive ncerts but no key, GnuTLS will
crash. GnuTLS should probably become robust to this scenario, just as it
must be robust to any manner of invalid input here. But it makes sense
for us to handle it, as well. We must abort the handshake in this case.

A new test is added that triggers this crash. Thanks to Martin Pitt for
contributing a reproducer that formed the basis for this test.

https://bugzilla.gnome.org/show_bug.cgi?id=793712