Move most of GTlsFileDatabaseGnutls up into GTlsDatabaseGnutls
Now that we use the GnuTLS system trust by default, it is likely that
the default GTlsDatabase is no longer backed by a file at all, since the
GnuTLS system trust uses the p11-kit trust store on Red Hat
distributions. But it *might* still be a file, which is still the case on
Debian, and when using the Freedesktop 1.6 SDK.

As suggested by Dan Winship, we can avoid this semantic difficulty by
moving pretty much all of the GTlsFileDatabaseGnutls implementation up
to GTlsDatabaseGnutls, and allowing instantiation of GTlsDatabaseGnutls.
GTlsDatabaseGnutls will now be used directly for the default
GTlsDatabase. The remaining code in GTlsFileDatabaseGnutls is just
enough to deal with the anchor_file property.

https://bugzilla.gnome.org/show_bug.cgi?id=753260