Don't generate installed test files when not enabled
These files are not needed except when installed tests are enabled, so
they should not be generated at all by default.

This is good to do for its own sake, but also works around some weird
behavior of meson that has resulted in these files being installed even
when installed tests are disabled.

https://bugzilla.gnome.org/show_bug.cgi?id=794372