Set a negative priority so glib-networking is loaded instead of glib-openssl
With this I am not saying that glib-openssl is less maintained
but that we want to use glib-networking instead of glib-openssl
for applications like Epiphany, since it is more tested for
those use cases. Instead I still encourage the usage of glib-openssl
since it is what we currently have available for windows and old
distributions.