Perform certificate verification during, not after, TLS handshake
We should perform certificate verification when GnuTLS expects us to do
so, during the TLS handshake, using
gnutls_session_set_verify_function(), rather than waiting until the
handshake has completed and then performing verification right after.
This is not currently a security problem, because we do perform
verification before any application data is sent, but this prepares the
codebase for TLS 1.3, and also happens to be a first step towards false
start for TLS 1.0-1.2, should we choose to enable that.

This is complicated by the accept-certificate signal. The verification
callback runs on the handshake thread, but the accept-certificate signal
needs to be emitted on the original (probably main) thread. So we keep
track of the thread-default main context at the time of the handshake.
In the case of sync handshakes, we have to iterate this main context in
a loop until the verification has been completed. This took me quite a
while to get right.

Fixes #6