tests: Allow insecure certificates in verify-with-incorrect-root-in-chain
This test checks that chain-with-verisign-md2.pem, which is not in the
trust store, is a valid substitute for ca-verisign-sha1.pem, since they
use the same private key. GnuTLS 3.6 now fails to validate this chain
since it no longer accepts MD2. This should be ignored for the purposes
of this test.

https://bugzilla.gnome.org/show_bug.cgi?id=794286