Disable building PKCS#11 backend by default
This backend has never been enabled at runtime, except when the
GIO_USE_TLS=gnutls-pkcs11 environment variable is set. I'm quite
uncertain as to whether it should continue to exist or not. So let's
disable it. If nobody complains, I'll probably delete the code, but
let's keep the code around for now to be on the safe side.

https://bugzilla.gnome.org/show_bug.cgi?id=793281