Fix leaks when calling g_tls_connection_gnutls_get_certificate
The out parameters are transfer full and need to be freed. This means we
have to store some member variables that would otherwise be unnecessary,
which is unfortunate, but we'll be able to get rid of them in the future
when switching to gnutls_certificate_set_retrieve_function3, once we
require GnuTLS 3.6.

Note that this is partially a regression from my commit c158d7f5, but
partially a preexisting issue. Originally we leaked 19 KB when running
the connection test. After my commit, we leaked 168 KB. With this
commit, there are no leaks at all.

Further note that we could have stored these members in the base class
and avoided some code duplication, but the base class is already way too
complex, and the code is more related to the child classes, and it seems
least-messy this way.

Thank you, asan!