Remove the three "during handshake" tests
Sadly, these tests rely on being able to run an async handshake on the
GTlsConnection without a running main context. That's no longer possible,
because verify_certificate_cb() now quite reasonably posts a call to
accept_certificate_cb() to the main context of the thread that started
the handshake. These tests use a thread-default main context that is not
iterated in order to ensure the handshake never finishes, which
previously was fine, but now the result is that the handshake doesn't
even get far enough for the accepted-cas property to be set, and so the
tests just iterate the default main context forever and time out.

I spent several hours trying to rewrite these tests, trying to use a
client authentication callback on the server to stall the handshake, but
was not successful, so let's just remove the tests. This isn't ideal,
as they are useful tests, but it is reasonable given that the tests are
now failing due to the way they are constructed, not due to an actual
bug.