Tighten up handling of server errors
A side effect of moving certificate verification inside the TLS
handshake is that the server now receives actual errors when the client
rejects its certificate. We should check for them. And we can't do so at
the natural end of the test case, because the server might not have
finished running yet, so there might not be any error yet. The right place
to do so is when shutting down the server.

This fixes #4, which was until now the only problematic test. (Without
this commit, a bunch of other tests would now also be racy.)

Finally, notice that the preexisting expect_server_error functionality
was not quite right, since it's only checking for errors closing the
connection. Turns out, that's all it was ever previously used for, and
the only test making use of it was one of the tests removed in the
previous commit, so that can go.