Fail unsafe rehandshake attempts initiated by API request
If the application attempts a rehandshake, that will currently be
allowed even if the server does not support safe renegotiation, and even
if the application has not enabled unsafe rehandshake mode. Since
applications have no way using our API to check if safe renegotiation is
supported, we had best check for them.