Remove unnecessary function parameter
start_async_server_and_connect_to_it() is always called with
should_close == TRUE. We can always add it back later if it's needed in
the future, but for now it's just harmful because it requires me to
think "what does TRUE mean?" when I see the callsites. It should become
an enum if it's needed again in the future.