Don't clear peer certificate when rehandshaking
The client auth tests check that the client connection's peer
certificate is nonnull, but it will in fact be null if the server has
initiated a rehandshake and the rehandshake has not yet finished. It
should be possible to avoid this by stalling the tests until the
rehandshake has completed, but there's really no need to clear it here
in the first place.

Of course, we still need to clear it when performing certificate
verification.