Fix connection failures caused by the certificate verification rework
I've been confused for a long time why 45c5f335 caused no problems to
our testsuite, but resulted in massive breakage when released in
2.57.90. Turns out the answer is session resumption. By moving
certificate verification into GnuTLS's certificate verification
callback, our manual verification was now not performed in cases of
session resumption. This means that the peer-certificate and
peer-certificate-errors properties were not updated properly. Our code
was not designed to cope with this.

So cope with it. We just have to manually update these properties. Be
super careful, because some of this code can now run on mulitple
threads (but hopefully not at the same time!). It's fine as long as we
make sure that application-visible notifies are only ever emitted on the
handshake context's thread, since application callbacks unexpectedly
running on secondary threads would be bad news. We went to a huge effort
to avoid that happening with the accept-certificate signal, so wouldn't
make sense to screw up notifies now.

Finally, clear handshake_context later in all codepaths, not for any
great technical reason, just so we can use it in
update_peer_certificate() to assert the function is called in the right
thread. Hopefully this shouldn't break anything.