Adapt client auth fail tests to TLS 1.3
There has been a surprising change is behavior regarding TLS client auth
failure with TLS 1.3, which I have documented here:

https://gitlab.com/gnutls/gnutls/issues/615

Basically the problem is that, if the client fails to send a required
certificate, or sends an unacceptable certificate, the handshake no
longer fails on the client side. This is a necessity of the protocol
changes, and it's awkward to attempt to hide this behind the GLib API.
So the errors we expect to receive in these failure tests are now
different than they used to be.

It's unclear to me how we will handle this change with other TLS
backends.