Remove the PKCS#11 backend
Nowadays the normal GnuTLS backend has full PKCS#11 support. (At least,
it's certainly supposed to.) The PKCS#11 backend has been disabled in
2.58, which is two months old now, and not a single complaint has
reached me from Ubuntu 18.10 or Fedora 29 or anywhere else. It's always
required a special environment variable to enable anyway, so I consider
all of this code experimental.

If anyone has a reason to keep this code around, please speak up now! We
can consider reverting this if there's a good reason to keep the PKCS#11
backend around that I'm not aware of. But I'm guessing there's not.

Fixes #7

Obsoletes #8