openssl: update SSL structure when certificate is changed
Set server certificate on SSL structure every time the certificate is
changed. Previously the certificate was set during initialization only.

This solves the problem when GTlsServerConnection is created without a
certificate and g_tls_connection_set_certificate() is used to supply the
certificate after creation time.

Also modify the connection unittest to cover this use case by setting
the server certificate after initialization when the default certificate
is used.

Fixes #55