.gitlab-ci.yml: Run tests 100x in a row
The glib-networking codebase is tricky and fragile, and the tests have a
history of flakiness. Currently, with DTLS disabled, I'm able to run the
tests 500x, which is incredible, but probably too much for CI. 100x
seems like a good middle ground. After a quick check, this is enough for
the DTLS tests to fail 3x, suggesting that 50x-100x is an appropriate
range and going too much lower risks possibly missing failures.

Also: somehow this works under asan without a single detected leak.
That's amazing.