Run tests 1000x
The timing issue here occurs too rarely on my computer, sometimes 5000
runs is enough to hit a failure but sometimes even 50000 runs is not
enough. It seems to be much flakier on CI so let's use CI to see if it's
fixed or not.

Goal is for this commit to fail CI with several failures, not just one
or two, so we can be confident the problem is fixed if the failures go
away.