Fix flaky failures in client auth failure tests
The test is currently too strict about which errors to expect after the
server drops the connection. It expects G_IO_ERROR_BROKEN_PIPE to only
occur during the close, but it could also occur during the read, in
which case there will be no error on close. It's a shame we can't
guarantee a consistent error in this case and the errors in general
could stand to be improved, but important thing is that there is some
error: the read is (properly) not working.

Fixes #66