base: Rename the streams to GTls[Input,Output]Stream
Instead of GTls[Input,Output]StreamBase, since these classes are not
inherited. This will be less confusing when the next commit turns them
into final types.

The naming is a bit iffy, since they are in the base/ directory after
all, but I think it will be less confusing this way.