database: use the default system store provided by openssl
And let the file database an independent class in case that we
need to use it.
Note the default database does not implement all the base class
methods since openssl does not provide us the api to iterate
over the ca certificates.
This is the reason the file database stays as is so we can iterate
over and have good unit tests. In theory for the file database
we should just add the file to the store and be done with it.