gnutls: Add missing calls to failed vfunc
This is needed wherever the connection can enter a failed state after a
handshake. We don't really have any way to track failed state, but it's
desirable to avoid connection reuse after errors.

It's non-obvious and not robust. We should try to improve this to make
it harder to break.