tests: Remove bad assertion from client-auth-failure test
This error needs to be checked when the server is closed. There's a
small but significant chance that might not have happened yet at this
point. We have already set expected_server_error to ensure the assert
happens, so this misplaced check is redundant and it's a simple mistake
that it wasn't removed already.