tests: Remove expected_server_error and manually wait instead
This effectively reverts 9e762996. It's required to solve a race in the
client auth failure test. We need to check for an error on the server
and then run the next test, but the expected_server_error functionality
isn't expressive enough to actually do that. So just manually add waits
everywhere instead, to make sure the server is finished, and then check
for the error. This is actually simpler than before anyway!