base: Don't initially cancel waiting_for_op
This is mostly harmless, but it causes a spurious wakeup the first time
a TLS source is created while an op is waiting. The cancellable source
fires immediately instead of when the op is yielded.

The cancellation here is not required for correctness because
tls_source_sync() will use a timeout source instead if there is no op or
IO waiting.