tests: test_client_auth_request_fail() should expect G_FILE_ERROR_ACCES
The FIXME is wrong. I misunderstood the purpose of this test when adding
the FIXME. The point of the test is to see what happens when the
GTlsInteraction's certificate request fails. Our mock request uses
G_FILE_ERROR_ACCES in this case, so that's the error we should expect.

We spent some time today on IRC trying to debug why we were getting
G_TLS_CERTIFICATE_ERROR instead for GnuTLS. Turns out, we're not. The
preprocessor guard is just misleading. OPENSSL_VERSION_NUMBER is not
defined when compiling for GnuTLS, so that turns into 0, and 0 <
0x10101000L, so we actually are building the first case here, not the
second. Whoops!