Fix openssl detection when there is no pkg-config file
Many SDK ship openssl headers and library without the .pc file, not only
on Windows. In that case we also need to fallback to find
the library and its headers.

Simplify the check for header files by using new syntax from Meson
0.50.0 that can search for a library and a set of headers in one call.

Fix find_library() calls by removing the 'lib' prefix, because Meson
already prepend it automatically, so it was searching for
bogus 'liblibcrypto.so'.